wip sdf rendering

+++ include/dr4/dr4_math.h
@@ -74,7 +74,7 @@ namespace dr4 {
 	inline float clampf(float val, float minval, float maxval){
 		return val < minval ? minval : (val > maxval ? maxval : val); 
 	}
-	inline float clampd(double val, double minval, double maxval){
+	inline double clampd(double val, double minval, double maxval){
 		return val < minval ? minval : (val > maxval ? maxval : val); 
 	}
